:lipstick: feat: adding token icon on modal
