+++ views/incinerator/incinerator-table/incinerator-table-body/incinerator-table-row.tsx
@@ -1,19 +1,13 @@
 import { Box, Checkbox, Typography } from '@interest-protocol/ui-kit';
-import { formatAddress } from '@mysten/sui.js/utils';
 import { FC } from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
 import { v4 } from 'uuid';
 
-import { TokenIcon } from '@/components';
-import { useNetwork } from '@/context/network';
 import { CoinObject } from '@/hooks/use-get-all-coins/use-get-all-coins.types';
 import { FixedPointMath } from '@/lib';
-import {
-  getKindFromObjectData,
-  getSymbolByType,
-  ZERO_BIG_NUMBER,
-} from '@/utils';
+import { ZERO_BIG_NUMBER } from '@/utils';
 
+import IncineratorTokenObject from '../../component/incinerator-token-object';
 import {
   IncineratorForm,
   IncineratorTableRowProps,
@@ -21,19 +15,9 @@ import {
 import QtyIncinerate from './incinerator-table-maybe-input';
 
 const IncineratorTableRow: FC<IncineratorTableRowProps> = ({ object }) => {
-  const network = useNetwork();
-  const { display, type } = object;
-  const displayName = display
-    ? (display as Record<string, string>).name ?? display.symbol ?? type
-    : type;
-
-  const { symbol, type: coinType } = (display as CoinObject) ?? {
-    type,
-    symbol: getSymbolByType(type),
-  };
+  const { display } = object;
 
   const index = object.index;
-  const url = (display as Record<string, string>)?.image_url;
   const { control, setValue } = useFormContext<IncineratorForm>();
   const active = useWatch({ control, name: `objects.${index}.active` });
   const editable = useWatch({ control, name: `objects.${index}.editable` });
@@ -44,8 +28,6 @@ const IncineratorTableRow: FC<IncineratorTableRowProps> = ({ object }) => {
   );
 
   const handleCheck = () => setValue(`objects.${index}.active`, !active);
-
-  const objectKind = getKindFromObjectData(object);
 
   return (
     <Box
@@ -66,36 +48,8 @@ const IncineratorTableRow: FC<IncineratorTableRowProps> = ({ object }) => {
       >
         <Checkbox defaultValue={active} onClick={handleCheck} label="" />
       </Typography>
-      <Typography
-        pr="m"
-        as="td"
-        gap="s"
-        py="xs"
-        size="small"
-        display="flex"
-        variant="label"
-        alignItems="center"
-      >
-        <TokenIcon
-          withBg
-          size="1.6rem"
-          symbol={symbol}
-          {...(url ? { url } : { type: coinType, network })}
-        />
-        <Box>
-          <Typography size="medium" variant="body" whiteSpace="nowrap">
-            {type === displayName ? formatAddress(type) : displayName}
-          </Typography>
-          <Typography
-            as="span"
-            size="small"
-            variant="body"
-            color="outline"
-            whiteSpace="nowrap"
-          >
-            Type: {objectKind}
-          </Typography>
-        </Box>
+      <Typography pr="m" as="td" py="xs" size="small" variant="label">
+        <IncineratorTokenObject object={object} />
       </Typography>
       <Typography
         pr="m"
